+++ spring-boot-exception-handler-starter/src/main/java/by/clevertec/exception/CustomEntityNotFoundException.java
@@ -1,4 +1,4 @@
-package by.clevertec.house.exception;
+package by.clevertec.exception;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.ResponseStatus;
