+++ spring-boot-exception-handler-starter/src/main/java/by/clevertec/model/BaseResponse.java
@@ -1,0 +1,16 @@
+package by.clevertec.model;
+
+import java.time.LocalDateTime;
+import java.time.format.DateTimeFormatter;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@Setter
+@AllArgsConstructor
+public abstract class BaseResponse {
+
+    private Integer status;
+    private final String timestamp = LocalDateTime.now().format(DateTimeFormatter.ISO_DATE_TIME);
+}
