+++ spring-boot-exception-handler-starter/src/main/resources/META-INF/spring/org.springframework.boot.autoconfigure.AutoConfiguration.imports
@@ -1,0 +1,1 @@
+by.clevertec.config.GlobalExceptionHandlerAutoConfiguration
