+++ src/main/java/by/clevertec/house/controller/PersonController.java
@@ -14,9 +14,9 @@ import by.clevertec.house.dto.request.PersonRequestDto;
 import by.clevertec.house.dto.response.HouseResponseDto;
 import by.clevertec.house.dto.response.PersonResponseDto;
 import by.clevertec.house.model.BaseResponse;
-import by.clevertec.house.model.ErrorValidationResponse;
-import by.clevertec.house.model.ExceptionResponse;
 import by.clevertec.house.service.PersonService;
+import by.clevertec.model.ErrorValidationResponse;
+import by.clevertec.model.ExceptionResponse;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
 import io.swagger.v3.oas.annotations.media.ArraySchema;
