+++ src/main/java/by/clevertec/house/service/impl/HouseServiceImpl.java
@@ -5,13 +5,13 @@ import static by.clevertec.house.util.ResponseUtils.CONDITIONAL_EXCEPTION_MESSAG
 import static by.clevertec.house.util.ResponseUtils.CONDITIONAL_HOUSE_OWNER_EXIST_EXCEPTION_MESSAGE;
 import static by.clevertec.house.util.ResponseUtils.CONDITIONAL_HOUSE_OWNER_NOT_EXIST_EXCEPTION_MESSAGE;
 
+import by.clevertec.exception.ConditionalException;
+import by.clevertec.exception.CustomEntityNotFoundException;
 import by.clevertec.house.domain.House;
 import by.clevertec.house.domain.Person;
 import by.clevertec.house.dto.request.HouseRequestDto;
 import by.clevertec.house.dto.response.HouseResponseDto;
 import by.clevertec.house.dto.response.PersonResponseDto;
-import by.clevertec.house.exception.ConditionalException;
-import by.clevertec.house.exception.CustomEntityNotFoundException;
 import by.clevertec.house.mapper.HouseMapper;
 import by.clevertec.house.mapper.PersonMapper;
 import by.clevertec.house.proxy.Cacheable;
