+++ src/test/java/by/clevertec/house/controller/HouseControllerTest.java
@@ -9,11 +9,11 @@ import static by.clevertec.house.util.ResponseUtils.DELETED_HOUSE_OWNER_MESSAGE;
 import static by.clevertec.house.util.ResponseUtils.DELETION_MESSAGE;
 import static by.clevertec.house.util.ResponseUtils.HTTP_NOT_READABLE_EXCEPTION_MESSAGE;
 import static by.clevertec.house.util.ResponseUtils.UPDATE_MESSAGE;
-import static by.clevertec.house.util.ResponseUtils.getExceptionResponse;
 import static by.clevertec.house.util.ResponseUtils.getSuccessResponse;
 import static by.clevertec.house.util.TestConstant.HOUSE_UUID;
 import static by.clevertec.house.util.TestConstant.INCORRECT_UUID;
 import static by.clevertec.house.util.TestConstant.PERSON_UUID;
+import static by.clevertec.util.ResponseUtils.getExceptionResponse;
 import static org.springframework.http.MediaType.APPLICATION_JSON;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.delete;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
@@ -23,6 +23,9 @@ import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.jsonPath;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
+import by.clevertec.exception.ConditionalException;
+import by.clevertec.exception.CustomEntityNotFoundException;
+import by.clevertec.exception.CustomNoContentException;
 import by.clevertec.house.AbstractTest;
 import by.clevertec.house.config.TestContainerConfig;
 import by.clevertec.house.domain.House;
@@ -30,14 +33,11 @@ import by.clevertec.house.domain.Person;
 import by.clevertec.house.dto.request.HouseRequestDto;
 import by.clevertec.house.dto.response.HouseResponseDto;
 import by.clevertec.house.dto.response.PersonResponseDto;
-import by.clevertec.house.exception.ConditionalException;
-import by.clevertec.house.exception.CustomEntityNotFoundException;
-import by.clevertec.house.exception.CustomNoContentException;
-import by.clevertec.house.model.ExceptionResponse;
 import by.clevertec.house.model.MessageResponse;
 import by.clevertec.house.proxy.HouseCacheableAspect;
 import by.clevertec.house.util.HouseTestBuilder;
 import by.clevertec.house.util.PersonTestBuilder;
+import by.clevertec.model.ExceptionResponse;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
