+++ src/test/java/by/clevertec/house/controller/HouseHistoryControllerTest.java
@@ -1,23 +1,23 @@
 package by.clevertec.house.controller;
 
-import static by.clevertec.house.util.ResponseUtils.getExceptionResponse;
 import static by.clevertec.house.util.TestConstant.HOUSE_UUID;
 import static by.clevertec.house.util.TestConstant.INCORRECT_UUID;
 import static by.clevertec.house.util.TestConstant.PERSON_UUID;
+import static by.clevertec.util.ResponseUtils.getExceptionResponse;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.jsonPath;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
+import by.clevertec.exception.CustomNoContentException;
 import by.clevertec.house.AbstractTest;
 import by.clevertec.house.config.TestContainerConfig;
 import by.clevertec.house.domain.House;
 import by.clevertec.house.domain.Person;
 import by.clevertec.house.dto.response.HouseResponseDto;
 import by.clevertec.house.dto.response.PersonResponseDto;
-import by.clevertec.house.exception.CustomNoContentException;
-import by.clevertec.house.model.ExceptionResponse;
 import by.clevertec.house.util.HouseTestBuilder;
 import by.clevertec.house.util.PersonTestBuilder;
+import by.clevertec.model.ExceptionResponse;
 import lombok.RequiredArgsConstructor;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
