+++ src/test/java/by/clevertec/house/controller/PersonControllerTest.java
@@ -7,12 +7,12 @@ import static by.clevertec.house.util.ResponseUtils.HTTP_NOT_READABLE_EXCEPTION_
 import static by.clevertec.house.util.ResponseUtils.METHOD_ARGUMENT_NOT_VALID_EXCEPTION_MESSAGE;
 import static by.clevertec.house.util.ResponseUtils.PERSON_CHANGE_HOUSE_MESSAGE;
 import static by.clevertec.house.util.ResponseUtils.UPDATE_MESSAGE;
-import static by.clevertec.house.util.ResponseUtils.getExceptionResponse;
 import static by.clevertec.house.util.ResponseUtils.getSuccessResponse;
 import static by.clevertec.house.util.TestConstant.ANOUTHER_HOUSE_UUID;
 import static by.clevertec.house.util.TestConstant.HOUSE_UUID;
 import static by.clevertec.house.util.TestConstant.INCORRECT_UUID;
 import static by.clevertec.house.util.TestConstant.PERSON_UUID;
+import static by.clevertec.util.ResponseUtils.getExceptionResponse;
 import static org.springframework.http.MediaType.APPLICATION_JSON;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.delete;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
@@ -23,6 +23,9 @@ import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.jsonPath;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
+import by.clevertec.exception.ConditionalException;
+import by.clevertec.exception.CustomEntityNotFoundException;
+import by.clevertec.exception.CustomNoContentException;
 import by.clevertec.house.AbstractTest;
 import by.clevertec.house.config.TestContainerConfig;
 import by.clevertec.house.domain.Gender;
@@ -33,15 +36,12 @@ import by.clevertec.house.dto.request.PersonPathRequestDto;
 import by.clevertec.house.dto.request.PersonRequestDto;
 import by.clevertec.house.dto.response.HouseResponseDto;
 import by.clevertec.house.dto.response.PersonResponseDto;
-import by.clevertec.house.exception.ConditionalException;
-import by.clevertec.house.exception.CustomEntityNotFoundException;
-import by.clevertec.house.exception.CustomNoContentException;
-import by.clevertec.house.model.ErrorValidationResponse;
-import by.clevertec.house.model.ExceptionResponse;
 import by.clevertec.house.model.MessageResponse;
 import by.clevertec.house.util.HouseTestBuilder;
 import by.clevertec.house.util.PassportTestBuilder;
 import by.clevertec.house.util.PersonTestBuilder;
+import by.clevertec.model.ErrorValidationResponse;
+import by.clevertec.model.ExceptionResponse;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
